change gold variable type to long

+++ src/UserInterface/AbstractPlayer.h
@@ -15,7 +15,7 @@ class IAbstractPlayer : public TAbstractSingleton<IAbstractPlayer>
 		virtual void	SetMainCharacterIndex(int iIndex) = 0;
 		virtual bool	IsMainCharacterIndex(DWORD dwIndex) = 0;
 
-		virtual int		GetStatus(DWORD dwType) = 0;
+		virtual long long		GetStatus(DWORD dwType) = 0;
 
 		virtual const char *	GetName() = 0;
